requirements.txt
requests
pytube3
requests_oauthlib
moviepy
